requirements.txt
aiofiles==23.1.0
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
Deprecated==1.2.14
distro==1.8.0
Django==4.1
django-crontab==0.7.1
exceptiongroup==1.1.3
frozenlist==1.4.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
magic-filter==1.0.12
multidict==6.0.4
netaddr==0.9.0
openai==1.2.3
pydantic==2.4.2
pydantic_core==2.10.1
pyTelegramBotAPI==4.14.0
python-dotenv==1.0.0
python-telegram-bot==20.6
pytz==2023.3.post1
requests==2.31.0
sniffio==1.3.0
sqlparse==0.4.4
telebot==0.0.5
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.7
wrapt==1.16.0
yarl==1.9.2
yookassa==2.5.0
